+++ bigquery_etl/backfill/validate.py
@@ -1,21 +1,21 @@
 """Validate backfill entries."""
-from ..backfill.parse_backfill import BackfillStatus
+from ..backfill.parse import Backfill, BackfillStatus
 
 
 def validate(backfill, backfills):
     """Check new backfill entry against existing backfills."""
     for entry_date, entry in backfills.items():
-        if entry.status == BackfillStatus.Drafting.value:
+        if entry.status == BackfillStatus.DRAFTING:
             validate_overlap_dates(backfill, entry)
 
 
 def validate_overlap_dates(entry_1: Backfill, entry_2: Backfill):
     """Check overlap dates between two backfill entries."""
-    if max(backfill.start_date, entry.start_date) <= min(
-        backfill.end_date, entry.end_date
+    if max(entry_1.start_date, entry_2.start_date) <= min(
+        entry_1.end_date, entry_2.end_date
     ):
         raise ValueError(
-            f"Existing backfill entry with overlap dates from: {entry.entry_date}."
+            f"Existing backfill entry with overlap dates from: {entry_1.entry_date}."
         )
